c2_profiles/dns_covert/requirements.txt
dnspython>=2.4.0
aiohttp>=3.8.0
cryptography>=41.0.0
